+++ src/Ds/Bundle/AccountBundle/Event/Listener/LocaleListener.php
@@ -3,7 +3,7 @@
 namespace Ds\Bundle\AccountBundle\Event\Listener;
 
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Translation\DataCollectorTranslator;
+use Symfony\Component\Translation\TranslatorInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 
 /**
@@ -17,7 +17,7 @@ class LocaleListener
     protected $stack;
 
     /**
-     * @var \Symfony\Component\Translation\DataCollectorTranslator
+     * @var \Symfony\Component\Translation\TranslatorInterface
      */
     protected $translator;
 
@@ -25,9 +25,9 @@ class LocaleListener
      * Constructor
      *
      * @param \Symfony\Component\HttpFoundation\RequestStack $stack
-     * @param \Symfony\Component\Translation\DataCollectorTranslator $translator
+     * @param \Symfony\Component\Translation\TranslatorInterface $translator
      */
-    public function __construct(RequestStack $stack, DataCollectorTranslator $translator)
+    public function __construct(RequestStack $stack, TranslatorInterface $translator)
     {
         $this->stack = $stack;
         $this->translator = $translator;
@@ -47,6 +47,7 @@ class LocaleListener
             return;
         }
 
-        $this->translator->setLocale($this->stack->getCurrentRequest()->getLocale());
+        $locale = $this->stack->getCurrentRequest()->getLocale();
+        $this->translator->setLocale($locale);
     }
 }