+++ src/Ds/Bundle/PortalBundle/CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog references the relevant changes (new features, changes and bugs) 
 
   * Added .codeclimate.yml file.
   * Fixed flatly-custom theme padding issue in navigation.
+  * Fixed locale listener translator argument type hinting.
 
 
 # CHANGELOG for 0.3.0